refactor: view RacingView 폴더화
